requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
sqlalchemy>=2.0.0

# HTTP clients
httpx>=0.27.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# CLI
click>=8.1.0
rich>=13.7.0

# Admin UI
flask>=3.0.0

# Utilities
tenacity>=8.2.0
